Basic Registration completed
Also updated packages
