+++ src/AppBundle/Controller/SecurityController.php
@@ -7,6 +7,8 @@ use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
+use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
+
 
 /**
  * Security controller.
@@ -103,42 +105,5 @@ class SecurityController extends Controller
 
       return $this->redirectToRoute('campaign_index', array('campaignUrl' => $campaign->getUrl()));
   }
-
-
-
-  /**
-   * @Route("/register", name="user_registration")
-   */
-  public function registerAction(Request $request, UserPasswordEncoderInterface $passwordEncoder)
-  {
-      // 1) build the form
-      $user = new User();
-      $form = $this->createForm(UserType::class, $user);
-
-      // 2) handle the submit (will only happen on POST)
-      $form->handleRequest($request);
-      if ($form->isSubmitted() && $form->isValid()) {
-
-          // 3) Encode the password (you could also do this via Doctrine listener)
-          $password = $passwordEncoder->encodePassword($user, $user->getPlainPassword());
-          $user->setPassword($password);
-
-          // 4) save the User!
-          $em = $this->getDoctrine()->getManager();
-          $em->persist($user);
-          $em->flush();
-
-          // ... do any other work - like sending them an email, etc
-          // maybe set a "flash" success message for the user
-
-          return $this->redirectToRoute('replace_with_some_route');
-      }
-
-      return $this->render(
-          'registration/register.html.twig',
-          array('form' => $form->createView())
-      );
-  }
-
 
 }
